Update in EnterActivity.kt way to check isLoggedIn
