+++ app/src/main/java/com/example/chessgo/frontend/mainmenu/MainMenuActivity.kt
@@ -1,76 +1,48 @@
 package com.example.chessgo.frontend.mainmenu
 
-import android.annotation.SuppressLint
-import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.Button
-import androidx.compose.material.Scaffold
-import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ExitToApp
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material.rememberScaffoldState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Button
-import androidx.compose.material.Text
-import androidx.compose.ui.unit.dp
-import com.example.chessgo.backend.global.UserAuthenticationData
-import com.example.chessgo.backend.registration.sign_in.SignInManager
 import com.example.chessgo.frontend.irlMenu.IRLMenuActivity
-import com.example.chessgo.frontend.mainmenu.onlinegame.OnlineGameMenuActivity
+import com.example.chessgo.frontend.onlinegame.OnlineGameMenuActivity
 import com.example.chessgo.frontend.registration.enterApp.EnterActivity
 import com.example.chessgo.ui.theme.ChessgoTheme
 import com.google.firebase.auth.FirebaseAuth
-import kotlinx.coroutines.launch
-import com.example.chessgo.ui.theme.ChessgoTheme
 
 class MainMenuActivity : ComponentActivity(){
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-    }
-
-    override fun onStart() {
-        super.onStart()
         setContent {
             ChessgoTheme {
                 MainScreen(
                     onSignOutClick = {
-                        val intent = Intent(applicationContext, EnterActivity::class.java)
-                        startActivity(intent)
+                        signOut()
+                        toEnterActivity()
                     },
-                    onIrlClick = {
-                        val intent = Intent(applicationContext, IRLMenuActivity::class.java)
-                        startActivity(intent)
-                    },
-                    onOnlineClick = {
-                        val intent = Intent(applicationContext, OnlineGameMenuActivity::class.java)
-                        startActivity(intent)
-                    },
-                    signOut = { signOut() })
+                    onIrlClick = { toIRLMenu() },
+                    onOnlineClick = { toOnlineMenu() }
+                )
             }
         }
     }
 
     fun signOut() {
-        UserAuthenticationData.removeAuthenticationData(context = this.baseContext)
         FirebaseAuth.getInstance().signOut()
+    }
+    fun toOnlineMenu() {
+        val intent = Intent(applicationContext, OnlineGameMenuActivity::class.java)
+        startActivity(intent)
+    }
+    fun toIRLMenu() {
+        val intent = Intent(applicationContext, IRLMenuActivity::class.java)
+        startActivity(intent)
+    }
+
+    fun toEnterActivity() {
+        val intent = Intent(applicationContext, EnterActivity::class.java)
+        startActivity(intent)
     }
 
     override fun onBackPressed() {
@@ -79,106 +51,4 @@ class MainMenuActivity : ComponentActivity(){
         // platform.
         return
     }
-}
-
-
-@SuppressLint("UnusedMaterialScaffoldPaddingParameter")
-@Composable
-fun MainScreen(
-    onSignOutClick: () -> Unit,
-    onIrlClick: () -> Unit,
-    onOnlineClick: () -> Unit,
-    signOut: () -> Unit
-) {
-
-    val scaffoldState = rememberScaffoldState()
-    val scope = rememberCoroutineScope()
-    Scaffold(
-        scaffoldState = scaffoldState,
-        topBar = {
-            AppBar(
-                onNavigationIconClick = {
-                    scope.launch {
-                        scaffoldState.drawerState.open()
-                    }
-                }
-            )
-        },
-        drawerGesturesEnabled = scaffoldState.drawerState.isOpen,
-        drawerContent = {
-            DrawerHeader(modifier = Modifier)
-            DrawerBody(
-                items = listOf(
-                    MenuItem(
-                        id = "home",
-                        title = "Home",
-                        contentDescription = "Go to home screen",
-                        icon = Icons.Default.Home
-                    ),
-                    MenuItem(
-                        id = "settings",
-                        title = "Settings",
-                        contentDescription = "Go to settings screen",
-                        icon = Icons.Default.Settings
-                    ),
-                    MenuItem(
-                        id = "help",
-                        title = "Help",
-                        contentDescription = "Get help",
-                        icon = Icons.Default.Info
-                    ),
-                    MenuItem(
-                        id = "signOut",
-                        title = "Sign Out",
-                        contentDescription = "signOut",
-                        icon = Icons.Default.ExitToApp
-                    ),
-                ),
-                onItemClick = {
-                    if (it.id == "signOut") {
-                        signOut()
-                        onSignOutClick()
-                    }
-                }
-            )
-        }
-    ) {
-        mainContent(
-            onIrlClick = onIrlClick,
-            onOnlineClick = onOnlineClick
-        )
-    }
-}
-
-@Composable
-fun mainContent(onIrlClick: () -> Unit, onOnlineClick: () -> Unit){
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(horizontal = 4.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-
-        ) {
-            Spacer(modifier = Modifier.height(20.dp))
-            Button(
-                onClick = { onIrlClick() },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(300.dp)
-                    .padding(vertical = 8.dp)
-            ) {
-                Text(text = "Play Irl")
-            }
-            Spacer(modifier = Modifier.height(20.dp))
-            Button(
-                onClick = { onOnlineClick() },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(300.dp)
-                    .padding(vertical = 8.dp)
-            ) {
-                Text(text = "Play Online")
-            }
-        }
 }
