+++ app/src/main/java/com/example/chessgo/frontend/mainmenu/MainMenuCompose.kt
@@ -1,0 +1,124 @@
+package com.example.chessgo.frontend.mainmenu
+
+import android.annotation.SuppressLint
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Button
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ExitToApp
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Info
+import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material.rememberScaffoldState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import kotlinx.coroutines.launch
+
+@SuppressLint("UnusedMaterialScaffoldPaddingParameter")
+@Composable
+fun MainScreen(
+    onSignOutClick: () -> Unit,
+    onIrlClick: () -> Unit,
+    onOnlineClick: () -> Unit
+) {
+
+    val scaffoldState = rememberScaffoldState()
+    val scope = rememberCoroutineScope()
+    Scaffold(
+        scaffoldState = scaffoldState,
+        topBar = {
+            AppBar(
+                onNavigationIconClick = {
+                    scope.launch {
+                        scaffoldState.drawerState.open()
+                    }
+                }
+            )
+        },
+        drawerGesturesEnabled = scaffoldState.drawerState.isOpen,
+        drawerContent = {
+            DrawerHeader(modifier = Modifier)
+            DrawerBody(
+                items = listOf(
+                    MenuItem(
+                        id = "home",
+                        title = "Home",
+                        contentDescription = "Go to home screen",
+                        icon = Icons.Default.Home
+                    ),
+                    MenuItem(
+                        id = "settings",
+                        title = "Settings",
+                        contentDescription = "Go to settings screen",
+                        icon = Icons.Default.Settings
+                    ),
+                    MenuItem(
+                        id = "help",
+                        title = "Help",
+                        contentDescription = "Get help",
+                        icon = Icons.Default.Info
+                    ),
+                    MenuItem(
+                        id = "signOut",
+                        title = "Sign Out",
+                        contentDescription = "signOut",
+                        icon = Icons.Default.ExitToApp
+                    ),
+                ),
+                onItemClick = {
+                    if (it.id == "signOut") {
+                        onSignOutClick()
+                    }
+                }
+            )
+        }
+    ) {
+        MainContent(
+            onIrlClick = onIrlClick,
+            onOnlineClick = onOnlineClick
+        )
+    }
+}
+
+@Composable
+fun MainContent(onIrlClick: () -> Unit, onOnlineClick: () -> Unit){
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(horizontal = 4.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center,
+
+        ) {
+        Spacer(modifier = Modifier.height(20.dp))
+        Button(
+            onClick = { onIrlClick() },
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(300.dp)
+                .padding(vertical = 8.dp)
+        ) {
+            Text(text = "Play Irl")
+        }
+        Spacer(modifier = Modifier.height(20.dp))
+        Button(
+            onClick = { onOnlineClick() },
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(300.dp)
+                .padding(vertical = 8.dp)
+        ) {
+            Text(text = "Play Online")
+        }
+    }
+}
